Better batching. Also added redraw function
Batching is now done globally and intelligently. It is slower
though so a performance test is needed and this commit could be the culprit
for lower performance. Assuming the amount of batchDraw calls in individual
nodes is limited the amortized performance shouldn't be too bad.

Also added the redraw function to all nodes! This function basically signals
that the node wants to be redrawn in the next frame. Every Node has a
clearBeforeDraw boolean property that decides if the draw should trigger a clear.
Setting this property to true causes the redraw call to chain upwards in the node
tree, with its final destination being the layer. If clearBeforeDraw is false,
the node queues a batchdraw of itself (and all its children) without clearing.

This way intermediary nodes can decide to improve drawing performance by
setting clearBeforeDraw to false if they know that this won't give artifacts.