Added boundsToPoints to Transform.
This function takes a bounds object, transforms the cornerpoints
of the bounding box and returns an array of the cornerpoints.

Is now used in calculateBoundingBox. Can also be used to quickly draw a
polygon that transforms along with the transformation of a node,
instead of only drawing a straight rectangle.

Example:

	var box = new Polygon({
		points: node.getTransform().boundsToPoints(node.calculateLocalBoundingBox()).flattenPoints(),
		stroke: 'black'
	});