Ignore detached nodes in delegated event handlers.
Fixes a long-standing 'bug' in delegated event handlers.
If one event handler detached or destroyed a node, the delegated
handlers that were checked after it (not necessarily handlers that would
trigger on the node in question) would crash because they tried to
walk its ancestor tree.

I've never been sure if this was the best place to implement this fix,
but I'm doing it now anyway.